+++ app/client/src/IDE/utils/getUpdatedTabs.ts
@@ -1,0 +1,7 @@
+export function getUpdatedTabs(newId: string, currentTabs: string[]) {
+  if (currentTabs.includes(newId)) return currentTabs;
+
+  const newTabs = [...currentTabs, newId];
+
+  return newTabs;
+}
